✨ 일지 결과 뼈 타워 구축
일지 결과 뼈 타워 아래붵 쌓아서 올라가도록 수정 minHeight 조정
